WORKING version of Algebra expand (!!!)
 -- This is a working version of the expand algebra function
 -- Givens an algebra expression e.g. (a+d)(b+f) it expands
    it into a DNF form WITHOUT using true tables.

This is significant because the DNF form via true tables
is 2^N were N is the number of terms i.e. very slow.
This is of order N^2 log(N) -- ie. for N==40 it is 1e8 times
faster!!!

There are an huge number of possible cases and I am not 100% I have
all of them working -- but I wrote a lot of tests in testAlgebra.
Anyone would be REALLY welcome to add a few more.