Reorder of Simulation::removeDeadSurfaces.
The removal of unused surfaces needs to take place both before
algebra optimization and afterwards. The algebra removal can remove
cells (which it did in this case) that are empty. This results in dead
surfaces that can be removed.

The original error was discovered by cell construction in
BremBlock/SquareFMask that produced empty cells. This construction
has also been supressed.